Split src/protein_quest/uniprot.py
