+++ src/protein_quest/mcp_server.py
@@ -48,6 +48,7 @@ from protein_quest.filters.ss import filter_file_on_secondary_structure
 from protein_quest.go import search_gene_ontology_term
 from protein_quest.io import convert_to_cif_file, glob_structure_files, read_structure
 from protein_quest.pdbe.fetch import fetch as pdbe_fetch
+from protein_quest.pdbe.result import PdbResult
 from protein_quest.structure import (
     nr_residues_in_chain,
     structure2uniprot_accessions,
@@ -55,7 +56,6 @@ from protein_quest.structure import (
 )
 from protein_quest.taxonomy import search_taxon
 from protein_quest.uniprot import (
-    PdbResult,
     Query,
     search4af,
     search4emdb,
